Support nested attribute names
This makes it possible to calculate histograms for nested values, e.g. 'meta.clinical.age'.

While this has no direct support in Resonant Laboratory yet, it was a feature introduced in the port to MSKCC that is being backported here for consistency between the codebases.

With this change, it might be possible to extract nested values from JSON data by manually adding the appropriate schema information to the dataset item's metadata.